Track log of component sizes

+++ hyracks-fullstack/hyracks/hyracks-storage-am-lsm-common/src/main/java/org/apache/hyracks/storage/am/lsm/common/api/ILSMOperationHistory.java
@@ -36,14 +36,14 @@ public class ILSMOperationHistory {
     public Semaphore mergeSem;
 
     // Pair<Pair<N, Size>, Average Latency>
-    public List<Pair<Pair<Integer, Long>, Double>> normalSearchHist;
+    public List<Pair<Pair<Integer, Double>, Double>> normalSearchHist;
     public int normalSearchCurrentN;
-    public long normalSearchCurrentSize;
+    public double normalSearchCurrentSize;
     public long normalSearchTotal;
     public long normalSearchCount;
-    public List<Pair<Pair<Integer, Long>, Double>> mergeSearchHist;
+    public List<Pair<Pair<Integer, Double>, Double>> mergeSearchHist;
     public int mergeSearchCurrentN;
-    public long mergeSearchCurrentSize;
+    public double mergeSearchCurrentSize;
     public long mergeSearchTotal;
     public long mergeSearchCount;
     public Semaphore searchSem;
@@ -146,7 +146,7 @@ public class ILSMOperationHistory {
     }
 
     public long finishMerge(int k, long mergedSize, long duration) {
-        double d = (double)duration / 1000000.0;
+        double d = (double) duration / 1000000.0;
         try {
             mergeSem.acquire();
             try {
@@ -183,16 +183,16 @@ public class ILSMOperationHistory {
     private void updateMergeSearchHistory() {
         if (mergeSearchHist.isEmpty()) {
             mergeSearchHist
-                    .add(new Pair<>(new Pair<>(new Integer(mergeSearchCurrentN), new Long(mergeSearchCurrentSize)),
+                    .add(new Pair<>(new Pair<>(new Integer(mergeSearchCurrentN), new Double(mergeSearchCurrentSize)),
                             computeAverageLatency(mergeSearchCount, mergeSearchTotal)));
             return;
         }
 
         int idx = -1;
         for (int i = 0; i < mergeSearchHist.size(); i++) {
-            Pair<Integer, Long> components = mergeSearchHist.get(i).getKey();
+            Pair<Integer, Double> components = mergeSearchHist.get(i).getKey();
             if (components.getKey().intValue() == mergeSearchCurrentN
-                    && components.getValue().longValue() == mergeSearchCurrentSize) {
+                    && components.getValue().doubleValue() == mergeSearchCurrentSize) {
                 idx = i;
                 break;
             }
@@ -202,7 +202,7 @@ public class ILSMOperationHistory {
             mergeSearchHist.remove(idx);
         }
 
-        mergeSearchHist.add(new Pair<>(new Pair<>(new Integer(mergeSearchCurrentN), new Long(mergeSearchCurrentSize)),
+        mergeSearchHist.add(new Pair<>(new Pair<>(new Integer(mergeSearchCurrentN), new Double(mergeSearchCurrentSize)),
                 computeAverageLatency(mergeSearchCount, mergeSearchTotal)));
 
         if (mergeSearchHist.size() == 101) {
@@ -213,16 +213,16 @@ public class ILSMOperationHistory {
     private void updateNormalSearchHistory() {
         if (normalSearchHist.isEmpty()) {
             normalSearchHist
-                    .add(new Pair<>(new Pair<>(new Integer(normalSearchCurrentN), new Long(normalSearchCurrentSize)),
+                    .add(new Pair<>(new Pair<>(new Integer(normalSearchCurrentN), new Double(normalSearchCurrentSize)),
                             computeAverageLatency(normalSearchCount, normalSearchTotal)));
             return;
         }
 
         int idx = -1;
         for (int i = 0; i < normalSearchHist.size(); i++) {
-            Pair<Integer, Long> components = normalSearchHist.get(i).getKey();
+            Pair<Integer, Double> components = normalSearchHist.get(i).getKey();
             if (components.getKey().intValue() == normalSearchCurrentN
-                    && components.getValue().longValue() == normalSearchCurrentSize) {
+                    && components.getValue().doubleValue() == normalSearchCurrentSize) {
                 idx = i;
                 break;
             }
@@ -233,7 +233,7 @@ public class ILSMOperationHistory {
         }
 
         normalSearchHist
-                .add(new Pair<>(new Pair<>(new Integer(normalSearchCurrentN), new Long(normalSearchCurrentSize)),
+                .add(new Pair<>(new Pair<>(new Integer(normalSearchCurrentN), new Double(normalSearchCurrentSize)),
                         computeAverageLatency(normalSearchCount, normalSearchTotal)));
 
         if (normalSearchHist.size() == 101) {
@@ -241,7 +241,7 @@ public class ILSMOperationHistory {
         }
     }
 
-    public void recordPointSearch(int n, long totalSize, long duration, Boolean duringMerge, Boolean inMemory) {
+    public void recordPointSearch(int n, double totalSize, long duration, Boolean duringMerge, Boolean inMemory) {
         try {
             searchSem.acquire();
             try {
@@ -369,8 +369,8 @@ public class ILSMOperationHistory {
                     if (n >= 2) {
                         double sumx12 = 0.0, sumx22 = 0.0, sumx1y = 0.0, sumx2y = 0.0, sumx1x2 = 0.0;
                         for (int i = 0; i < n; i++) {
-                            Pair<Pair<Integer, Long>, Double> p = normalSearchHist.get(i);
-                            Pair<Integer, Long> components = p.getKey();
+                            Pair<Pair<Integer, Double>, Double> p = normalSearchHist.get(i);
+                            Pair<Integer, Double> components = p.getKey();
                             int x1 = components.getKey().intValue();
                             double x2 = components.getValue().doubleValue();
                             double y = p.getValue().doubleValue();
@@ -426,8 +426,8 @@ public class ILSMOperationHistory {
                     if (n >= 2) {
                         double sumx12 = 0.0, sumx22 = 0.0, sumx1y = 0.0, sumx2y = 0.0, sumx1x2 = 0.0;
                         for (int i = 0; i < n; i++) {
-                            Pair<Pair<Integer, Long>, Double> p = mergeSearchHist.get(i);
-                            Pair<Integer, Long> components = p.getKey();
+                            Pair<Pair<Integer, Double>, Double> p = mergeSearchHist.get(i);
+                            Pair<Integer, Double> components = p.getKey();
                             int x1 = components.getKey().intValue();
                             double x2 = components.getValue().doubleValue();
                             double y = p.getValue().doubleValue();
@@ -502,12 +502,12 @@ public class ILSMOperationHistory {
         return "[" + ret + "]";
     }
 
-    private static String searchHistToString(List<Pair<Pair<Integer, Long>, Double>> hist, int n, long size, long total,
-            long count) {
+    private static String searchHistToString(List<Pair<Pair<Integer, Double>, Double>> hist, int n, double size,
+            long total, long count) {
         String ret = "";
         for (int i = 0; i < hist.size(); i++) {
-            Pair<Pair<Integer, Long>, Double> p = hist.get(i);
-            Pair<Integer, Long> components = p.getKey();
+            Pair<Pair<Integer, Double>, Double> p = hist.get(i);
+            Pair<Integer, Double> components = p.getKey();
             if (i == 0) {
                 ret = "(" + components.getKey() + "," + components.getValue() + ")=" + p.getValue();
             } else {
